--- packages/utils/src/portSelectorFn.ts
@@ -1,20 +1,0 @@
-import { Logger } from '@graphql-mesh/types';
-
-export function portSelectorFn(sources: [number, number, number], logger: Logger) {
-  const port = sources.find(source => Boolean(source)) || 4000;
-  if (sources.filter(source => Boolean(source)).length > 1) {
-    const activeSources: Array<string> = [];
-    if (sources[0]) {
-      activeSources.push('CLI');
-    }
-    if (sources[1]) {
-      activeSources.push('serve configuration');
-    }
-    if (sources[2]) {
-      activeSources.push('environment variable');
-    }
-    logger.warn(`Multiple ports specified (${activeSources.join(', ')}), using ${port}`);
-  }
-
-  return port;
-}
